Simplified parameter names by removing _id prefix
When I added binary_get() and binary_set() I used _id prefixes after the
item and key parameters. This didn't line up with what was being used in
the existing functions. The function doc also didn't even match the
names of the parameters, so they have been updated as well.