Binary API: Renamed optional parameter to binary_get()
binary_get() should return binary data - bytes. If someone wants
to get the URL to the data instead of the data itself, they ought to
specifically request that. Therefore, instead of making the parameter
name fetch_data=True, I changed it to fetch_url=False, keeping
the default option to returning bytes, but making the parameter more
self-documenting on what it does.