Binary API: Reverted changes made to Bucket.get() and Bucket.set()
Initially, when adding binary_get() and binary_set(), get() and set()
were modified to take an option key parameter, which would wrap around
to calls to binary_get() and binary_set() with the same item and key.
This is redundant because the library users will already be making a
distinction between saving simple data verses binary data and can make
the direct call to binary_get() and binary_set().