Binary API: Give an extra hour to work with data URLs before they expire.
The BinaryResponse that is returned from a call to binary_get() has
a built-in URL validation check that will raise a UrlError if someone
tries to access it after its predetermined expiration time.

Before, this validation check compared the expiration time to the
instantaneous time of execution of the code. This could open up
situations where the user requests the data and the validation check
asserts valid, but where the end user might have only a second or two
to use that access before it does expire.

Because of this, we now raise the error an hour before the actual URL
expiration to give the end-developers a little extra time to access the
URLs or the data.

One (1) hour is an arbitrary amount of time chosen because it's
sufficiently long enough that no network/processing lag will interfere
but short enough that we don't expect too-eager of pre-expiration.